주식 목록 등락률 표시 오류 (#68)
* 주식 등락률은 실수형으로 표시하기로 했어서, 수정함.

* 주식 목록 - 등락률 소수점으로 표시되게 수정

+++ src/com/school/stockGame/dao/StockDetailDAO.java
@@ -290,11 +290,11 @@ public class StockDetailDAO {
 	}
 
 	// 주식 등락률 조회
-	public int getChangeRate(int StockNo) {
+	public double getChangeRate(int StockNo) {
 		Connection conn = null;
 		PreparedStatement stmt = null;
 		ResultSet rs = null;
-		int percent = 0;
+		double percent = 0.0;
 		try {
 			conn = DBCP.getConnection();
 			stmt = conn.prepareStatement(StockDetailQuery.STOCK_CHANGE_RATE_SQL);
